+++ packages/implementations/react/.storybook/main.ts
@@ -1,4 +1,4 @@
-import main from ".storybook/main";
+import main from "../../../../.storybook/main";
 import type { StorybookConfig } from "@storybook/react-webpack5";
 
 const config: StorybookConfig = {