+++ packages/implementations/react/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
+## 2024-04-18 `1.11.0`
+
+#### 🎉 New features
+
+- Added new atomic `Toggle` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-04-18 `1.10.0`
 
 #### 🎉 New features